shopify/requirements.txt
requests==2.32.3
mysql-connector-python==8.0.33 # Or a more recent stable version
